refactor: move validation interface to protocols folder
