+++ src/presentation/helpers/validators/index.ts
@@ -1,4 +1,4 @@
-export * from './validation'
+export * from '../../protocols/validation'
 export * from './validation-composite'
 export * from './required-field'
 export * from './compare-fields'
