+++ src/presentation/helpers/validators/required-field/index.ts
@@ -1,5 +1,5 @@
 import { NoProvidedParamError } from '@/presentation/errors'
-import { IValidation } from '../validation'
+import { IValidation } from '../../../protocols/validation'
 
 export class RequiredFieldValidation implements IValidation {
   private readonly fieldName: string
