+++ src/presentation/helpers/validators/validation-composite/index.spec.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest'
 
 import { NoProvidedParamError } from '@/presentation/errors'
 import { ValidationComposite } from '.'
-import { IValidation } from '../validation'
+import { IValidation } from '../../../protocols/validation'
 
 const makeValidation = (): IValidation => {
   class ValidationStub implements IValidation {
