+++ src/presentation/helpers/validators/validation-composite/index.ts
@@ -1,4 +1,4 @@
-import { IValidation } from '../validation'
+import { IValidation } from '../../../protocols/validation'
 
 export class ValidationComposite implements IValidation {
   private readonly validations: IValidation[]
